+++ src/Contracts/Toolkit/Iterables/ListIterator.php
@@ -16,7 +16,7 @@ use IteratorAggregate;
 
 /**
  * @template TValue
- * @extends IteratorAggregate<TValue>
+ * @extends IteratorAggregate<int, TValue>
  */
 interface ListIterator extends IteratorAggregate, Countable
 {