+++ src/Toolkit/Pipeline/PipelineBuilder.php
@@ -83,7 +83,7 @@ final class PipelineBuilder implements IPipelineBuilder
             return $stage;
         }
 
-        if (is_string($stage) && $this->container) {
+        if ($this->container) {
             return new LazyPipe($this->container, $stage);
         }