+++ src/Toolkit/Result/KeyedSetOfErrors.php
@@ -85,7 +85,6 @@ final class KeyedSetOfErrors implements KeyedSet
         $copy = clone $this;
 
         foreach (self::from($other) as $key => $errors) {
-            assert(is_string($key) && $errors instanceof IListOfErrors);
             $copy->stack[$key] = $copy->get($key)->merge($errors);
         }