+++ tests/Unit/Toolkit/ContractsTest.php
@@ -23,6 +23,7 @@ class ContractsTest extends TestCase
     public function testItDoesNotThrowWhenPreconditionIsTrue(): void
     {
         Contracts::assert(true, 'Not expected error.');
+        /** @phpstan-ignore-next-line */
         $this->assertTrue(true);
     }