+++ tests/Unit/Toolkit/Identifiers/PossiblyNumericIdTest.php
@@ -42,11 +42,7 @@ class PossiblyNumericIdTest extends TestCase
     public function test(string|int $value, string|int $expected): void
     {
         $actual = new PossiblyNumericId($value);
-        $expectedId = match (true) {
-            is_string($expected) => new StringId($expected),
-            is_int($expected) => new IntegerId($expected),
-            default => $this->fail('Unexpected value.'),
-        };
+        $expectedId = is_string($expected) ? new StringId($expected) : new IntegerId($expected);
 
         $this->assertSame($expected, $actual->value);
         $this->assertSame($expected, PossiblyNumericId::from($value)->value);