+++ tests/Unit/Toolkit/Iterables/IsLazyListTest.php
@@ -15,7 +15,7 @@ use CloudCreativity\Modules\Contracts\Toolkit\Iterables\LazyList;
 use CloudCreativity\Modules\Toolkit\Iterables\IsLazyList;
 use PHPUnit\Framework\TestCase;
 
-class LazyListTraitTest extends TestCase
+class IsLazyListTest extends TestCase
 {
     /**
      * @return void